fix: nits, port context menu functions into own file
